+++ mockserver-core/src/main/java/org/mockserver/responsewriter/ResponseWriter.java
@@ -3,15 +3,19 @@ package org.mockserver.responsewriter;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import org.mockserver.configuration.Configuration;
 import org.mockserver.cors.CORSHeaders;
+import org.mockserver.log.model.LogEntry;
+import org.mockserver.logging.MockServerLogger;
 import org.mockserver.model.ConnectionOptions;
 import org.mockserver.model.HttpRequest;
 import org.mockserver.model.HttpResponse;
 import org.mockserver.version.Version;
+import org.slf4j.event.Level;
 
-import static io.netty.handler.codec.http.HttpHeaderNames.CONNECTION;
-import static io.netty.handler.codec.http.HttpHeaderNames.CONTENT_TYPE;
+import static io.netty.handler.codec.http.HttpHeaderNames.*;
 import static io.netty.handler.codec.http.HttpHeaderValues.CLOSE;
 import static io.netty.handler.codec.http.HttpHeaderValues.KEEP_ALIVE;
+import static org.apache.commons.lang3.StringUtils.isNotBlank;
+import static org.mockserver.log.model.LogEntry.LogMessageType.INFO;
 import static org.mockserver.mock.HttpState.PATH_PREFIX;
 import static org.mockserver.model.Header.header;
 import static org.mockserver.model.HttpResponse.notFoundResponse;
@@ -23,11 +27,13 @@ import static org.mockserver.model.HttpResponse.response;
 public abstract class ResponseWriter {
 
     protected final Configuration configuration;
+    protected final MockServerLogger mockServerLogger;
     private final CORSHeaders corsHeaders;
 
-    protected ResponseWriter(Configuration configuration) {
-        corsHeaders = new CORSHeaders(configuration);
+    protected ResponseWriter(Configuration configuration, MockServerLogger mockServerLogger) {
         this.configuration = configuration;
+        this.mockServerLogger = mockServerLogger;
+        corsHeaders = new CORSHeaders(configuration);
     }
 
     public void writeResponse(final HttpRequest request, final HttpResponseStatus responseStatus) {
@@ -53,6 +59,25 @@ public abstract class ResponseWriter {
             corsHeaders.addCORSHeaders(request, response);
         } else if (apiResponse && configuration.enableCORSForAPI()) {
             corsHeaders.addCORSHeaders(request, response);
+        }
+        String contentLengthHeader = response.getFirstHeader(CONTENT_LENGTH.toString());
+        if (isNotBlank(contentLengthHeader)) {
+            try {
+                int contentLength = Integer.parseInt(contentLengthHeader);
+                if (response.getBodyAsRawBytes().length > contentLength) {
+                    mockServerLogger.logEvent(
+                        new LogEntry()
+                            .setType(INFO)
+                            .setLogLevel(Level.INFO)
+                            .setCorrelationId(request.getLogCorrelationId())
+                            .setHttpRequest(request)
+                            .setHttpResponse(response)
+                            .setMessageFormat("returning response with content-length header " + contentLength + " which is smaller then response body length " + response.getBodyAsRawBytes().length + ", body will likely be truncated by client receiving request")
+                    );
+                }
+            } catch (NumberFormatException ignore) {
+                // ignore exception while parsing invalid content-length header
+            }
         }
         if (apiResponse) {
             response.withHeader("version", Version.getVersion());